Extract authentication base service class
